// launchdarkly/types/paths/api/v2/projects/{projectKey}/environments/{environmentKey}.types.ts
export type getEnvironment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: getEnvironment_Path;
    headers: getEnvironment_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Environment;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: ForbiddenErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type patchEnvironment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: patchEnvironment_Path;
    headers: patchEnvironment_Headers;
    cookie: never;
    body: JSONPatch;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Environment;
          };
        };
        examples: {};
      };
      400: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: InvalidRequestErrorRep;
          };
        };
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      409: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: StatusConflictErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Delete a environment by key.
 */
export type deleteEnvironment = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: deleteEnvironment_Path;
    headers: deleteEnvironment_Headers;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: UnauthorizedErrorRep;
          };
        };
        examples: {};
      };
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: ForbiddenErrorRep;
          };
        };
        examples: {};
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: NotFoundErrorRep;
          };
        };
        examples: {};
      };
      429: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: RateLimitedErrorRep;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    auth: { apiKey: string };
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type getEnvironment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
};

export type getEnvironment_Headers = { Authorization: string };

export type patchEnvironment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
};

export type patchEnvironment_Headers = { Authorization: string };

export type deleteEnvironment_Path = {
  /**
   * The project key
   */
  projectKey: string;
  /**
   * The environment key
   */
  environmentKey: string;
};

export type deleteEnvironment_Headers = { Authorization: string };